renamed to their class name
